Rename ui/reader package to ui/viewer
Renamed `ui/reader` directory to `ui/viewer` and updated package declarations in all affected Activity classes.
Updated references in imports, manifests, and layout files to point to the new package.
This change consolidates viewer-related activities into a more appropriately named package.
