+++ app/src/main/AndroidManifest.xml
@@ -89,7 +89,7 @@
         <activity
             android:name=".ui.resources.AddResourceActivity"
             android:theme="@style/AppTheme" />
-        <activity android:name=".ui.reader.AudioPlayerActivity"
+        <activity android:name=".ui.viewer.AudioPlayerActivity"
             android:configChanges="orientation|screenSize" />
         <activity
             android:name=".ui.feedback.FeedbackDetailActivity"
@@ -100,19 +100,19 @@
             android:label="@string/title_activity_dashboard"
             android:windowSoftInputMode="adjustResize"
             android:theme="@style/FullscreenTheme" />
-        <activity android:name=".ui.reader.PDFReaderActivity"
+        <activity android:name=".ui.viewer.PDFReaderActivity"
             android:configChanges="orientation|screenSize" />
-        <activity android:name=".ui.reader.WebViewActivity"
+        <activity android:name=".ui.viewer.WebViewActivity"
             android:configChanges="orientation|screenSize" />
-        <activity android:name=".ui.reader.VideoViewerActivity"
+        <activity android:name=".ui.viewer.VideoViewerActivity"
             android:configChanges="orientation|screenSize" />
-        <activity android:name=".ui.reader.ImageViewerActivity"
+        <activity android:name=".ui.viewer.ImageViewerActivity"
             android:configChanges="orientation|screenSize" />
-        <activity android:name=".ui.reader.TextFileViewerActivity"
+        <activity android:name=".ui.viewer.TextFileViewerActivity"
             android:configChanges="orientation|screenSize" />
-        <activity android:name=".ui.reader.MarkdownViewerActivity"
+        <activity android:name=".ui.viewer.MarkdownViewerActivity"
             android:configChanges="orientation|screenSize" />
-        <activity android:name=".ui.reader.CSVViewerActivity"
+        <activity android:name=".ui.viewer.CSVViewerActivity"
             android:configChanges="orientation|screenSize" />
         <activity
             android:name=".ui.dictionary.DictionaryActivity"
