+++ app/src/main/java/org/ole/planet/myplanet/utilities/ResourceOpener.kt
@@ -9,13 +9,13 @@ import org.ole.planet.myplanet.R
 import org.ole.planet.myplanet.base.BaseResourceFragment
 import org.ole.planet.myplanet.model.RealmMyLibrary
 import org.ole.planet.myplanet.service.UserSessionManager
-import org.ole.planet.myplanet.ui.reader.AudioPlayerActivity
-import org.ole.planet.myplanet.ui.reader.CSVViewerActivity
-import org.ole.planet.myplanet.ui.reader.ImageViewerActivity
-import org.ole.planet.myplanet.ui.reader.MarkdownViewerActivity
-import org.ole.planet.myplanet.ui.reader.PDFReaderActivity
-import org.ole.planet.myplanet.ui.reader.TextFileViewerActivity
-import org.ole.planet.myplanet.ui.reader.VideoViewerActivity
+import org.ole.planet.myplanet.ui.viewer.AudioPlayerActivity
+import org.ole.planet.myplanet.ui.viewer.CSVViewerActivity
+import org.ole.planet.myplanet.ui.viewer.ImageViewerActivity
+import org.ole.planet.myplanet.ui.viewer.MarkdownViewerActivity
+import org.ole.planet.myplanet.ui.viewer.PDFReaderActivity
+import org.ole.planet.myplanet.ui.viewer.TextFileViewerActivity
+import org.ole.planet.myplanet.ui.viewer.VideoViewerActivity
 
 object ResourceOpener {
     private fun resourcePath(item: RealmMyLibrary): String {
