+++ app/src/main/res/layout/activity_exo_player_video.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="org.ole.planet.myplanet.ui.reader.VideoViewerActivity">
+    tools:context="org.ole.planet.myplanet.ui.viewer.VideoViewerActivity">
 
     <androidx.media3.ui.PlayerView
         android:id="@+id/exo_player_simple"
