+++ app/src/main/res/layout/activity_image_viewer.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context="org.ole.planet.myplanet.ui.reader.ImageViewerActivity">
+    tools:context="org.ole.planet.myplanet.ui.viewer.ImageViewerActivity">
 
     <TextView
         android:id="@+id/imageFileName"
