+++ app/src/main/res/layout/activity_pdfreader.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context="org.ole.planet.myplanet.ui.reader.PDFReaderActivity">
+    tools:context="org.ole.planet.myplanet.ui.viewer.PDFReaderActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
