+++ app/src/main/res/layout/activity_textfile_viewer.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context="org.ole.planet.myplanet.ui.reader.TextFileViewerActivity">
+    tools:context="org.ole.planet.myplanet.ui.viewer.TextFileViewerActivity">
 
     <TextView
         android:id="@+id/textFileName"
